refactor jet sequences (part 2)
add configs that alias jet producers to the 'nominal' labels expected by
other producers and analysers. only one of any type of configs should be
imported.

this is so that every jet producer is defined in one place, with the
intended settings, instead of having multiple definitions or configs
that sneakily modify some parameters.